presentation-api/controlling-ua/defaultRequest_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API, testing to start a new presentation with an default request setup. (success - manual)</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p>Click the button or the menu item for presentation on your browser (for example, "Cast"), to start the manual test.</p>
<p id="notice">If your browser does not support <code>defaultRequest</code>, please click this button: <button id="notsupported">Not Supported</button>

<script>
    // disable the timeout function for the tests
    // and call 'done()' when the tests cases are finished.
    setup({explicit_timeout: true});

    // -----------
    // DOM Element
    // -----------
    var button = document.getElementById('notsupported'),
        notice = document.getElementById('notice');

    // ------------------------------
    // Start New Presentation with
    // 'default request' Test - BEGIN
    // ------------------------------
    async_test(function(t) {
        var request = new PresentationRequest(presentationUrls);
        if('defaultRequest' in navigator.presentation) {
            navigator.presentation.defaultRequest = request;

            if('onconnectionavailable' in request) {
                notice.parentNode.removeChild(notice);
                request.onconnectionavailable = t.step_func_done(function (evt) {
                    var connection = evt.connection;

                    assert_equals(connection.state, 'connecting', 'The initial state of the presentation connection is "connecting".');
                    assert_true(!connection.id, 'The connection ID is set.');
                    assert_true(typeof connection.id === 'string', 'The connection ID is a string.');
                    assert_true(connection instanceof PresentationConnection, 'The connection is an instance of PresentationConnection.');
                });
                button.onclick = t.step_func_done(function() {
                    button.disabled = true;
                    assert_unreached('This browser does not support defaultRequest.');
                });
            }
            else {
                notice.parentNode.removeChild(notice);
                assert_unreached('This browser does not support connectionavailable event.');
            }
        }
        else {
            notice.parentNode.removeChild(notice);
            assert_unreached('This browser does not support defaultRequest.');
        }
    }, '[Optional] The presentation was started successfully.');
    // ----------------------------
    // Start New Presentation with
    // 'default request' Test - END
    // ----------------------------
</script>
